+++ src/lib/types.ts
@@ -1,0 +1,13 @@
+export type FieldType = 'text' | 'number' | 'text array';
+
+export interface Field {
+  type: FieldType;
+  label: string;
+  id: string;
+}
+
+export interface Template {
+  id: string;
+  name: string;
+  content: string;
+}
